+++ inst/shiny-examples/OAuth/ui.R
@@ -1,0 +1,5 @@
+library(shiny)
+
+shinyUI(
+    uiOutput("ui")
+)