+++ NEWS.md
@@ -1,3 +1,32 @@
+# datimutils 0.6.0
+
+## New features
+* Adds `loginToDATIMOAuth` function to assist with OAUTH2.0 authentication. 
+
+# datimutils 0.5.4
+
+## Bug fixes
+* Fixes error conditions in getDataValueSets.
+
+
+# datimutils 0.5.2
+
+## Bug fixes
+* Fixes is timeout error handling in API calls.
+
+
+# datimutils 0.5.1
+
+## New features
+* Adds `verbose` parameter to all functions which returns entire response. Set to `FALSE` by default.
+* Adds `quiet` parameter to all functions which determines whether URL is printed. Set to `TRUE` by default.
+
+# datimutils 0.5.0
+
+## New features
+* Adds `getDataStoreKey` (DP-533)
+* Adds `convertFYQToCQ` (DP-132)
+
 # datimutils 0.4.0
 
 ## New features